+++ test/cram/team-lint.t/run.t
@@ -1,7 +1,7 @@
 
 Team Lint example
 
-  $ okra team lint -C admin/ -W 40-42 -y 2022 --conf ./conf.yml
+  $ okra team lint -C admin/ -w 40-42 -y 2022 --conf ./conf.yml
   === My Team ===
     + Engineer 1
       + Report week 40: Complete
@@ -17,6 +17,42 @@ Team Lint example
 
 Missing [--repo-dir] argument:
 
-  $ okra team lint -W 40-42 -y 2022 --conf ./conf.yml
+  $ okra team lint -w 40-42 -y 2022 --conf ./conf.yml
   Missing [-C] or [--repo-dir] argument, or [admin_dir] configuration.
   [1]
+
+Wrong month:
+
+  $ okra team lint -C admin/ -m 0 -y 2022 --conf ./conf.yml
+  invalid month: 0
+  [1]
+  $ okra team lint -C admin/ -m 13 -y 2022 --conf ./conf.yml
+  invalid month: 13
+  [1]
+
+Wrong week
+
+  $ okra team lint -C admin/ -w 53 -y 2019 --conf ./conf.yml
+  invalid week: 53
+  [1]
+  $ okra team lint -C admin/ -w 53 -y 2023 --conf ./conf.yml
+  invalid week: 53
+  [1]
+  $ okra team lint -C admin/ -w 53 -y 2024 --conf ./conf.yml
+  invalid week: 53
+  [1]
+  $ okra team lint -C admin/ -w 0 -y 2024 --conf ./conf.yml
+  invalid week: 0
+  [1]
+  $ okra team lint -C admin/ -w 2-1 -y 2024 --conf ./conf.yml
+  invalid week range: 2-1
+  [1]
+
+2020 has 53 weeks:
+
+  $ okra team lint -C admin/ -w 53 -y 2020 --conf ./conf.yml
+  === My Team ===
+    + Engineer 1
+      + Report week 53: Not found: admin//weekly/2020/53/eng1.md
+    + Engineer 2
+      + Report week 53: Not found: admin//weekly/2020/53/eng2.md